Create more specific files for constants
