+++ package.json
@@ -8,7 +8,7 @@
   "private": true,
   "engines": {
     "npm": ">4",
-    "node": ">10.16"
+    "node": ">=10"
   },
   "scripts": {
     "start": "NODE_ENV=development start-storybook -s ./public -p 9001 -c .storybook",
