+++ src/storybook/foundation/utilities/display.js
@@ -1,0 +1,17 @@
+export const displayClasses = {
+  display: ['.d-{value}', '.d-{breakpoint}-{value}'],
+  values: [
+    'none',
+    'inline',
+    'inline-block',
+    'block',
+    'table',
+    'cell',
+    'table-cell',
+    'table-row',
+    'flex',
+    'inline-flex'
+  ]
+}
+
+export default displayClasses
