+++ src/storybook/foundation/utilities/flex.js
@@ -1,4 +1,4 @@
-export const flexClasses = {
+const flexClasses = {
   display: [
     '.d-flex',
     '.d-inline-flex',
@@ -86,18 +86,4 @@ export const flexClasses = {
   ]
 }
 
-export const displayClasses = {
-  display: ['.d-{value}', '.d-{breakpoint}-{value}'],
-  values: [
-    'none',
-    'inline',
-    'inline-block',
-    'block',
-    'table',
-    'cell',
-    'table-cell',
-    'table-row',
-    'flex',
-    'inline-flex'
-  ]
-}
+export default flexClasses
